requirements.txt
Django>=2.0
django-rest-framework
psycopg2-binary
